bootimage: create marketplace pkg
